- (config*, agent/mibgroup/*.h):   - new config macro:  config_arch_require(arch, module).   - moved arch dependant modules (memory, hpux, kernel_sunos5) loads     from the configure into the appropriate .h files. - (interfaces.c, var_route.c):   - typecast klookup calls for freebsd2.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@979 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/interfaces.c
@@ -1176,7 +1176,7 @@ struct in_ifaddr *Retin_ifaddr;
           /*
            *      Get the "in_ifaddr" structure
            */
-          klookup(in_ifaddraddr, (char *)&in_ifaddr, sizeof in_ifaddr);
+          klookup((unsigned long)in_ifaddraddr, (char *)&in_ifaddr, sizeof in_ifaddr);
           in_ifaddraddr = in_ifaddr.ia_next;
 
           if (Retin_ifaddr)
@@ -1189,7 +1189,7 @@ struct in_ifaddr *Retin_ifaddr;
 
               KNLookup (interfaces_nl, N_IFNET, (char *)&ifnetaddr, sizeof(ifnetaddr));
               while (ifnetaddr && ifnetaddr != in_ifaddr.ia_ifp) {
-                      klookup(ifnetaddr, (char *)&ifnet, sizeof ifnet);
+                      klookup((unsigned long)ifnetaddr, (char *)&ifnet, sizeof ifnet);
                       ifnetaddr = ifnet.if_next;
                       index++;
               }
